feat(vehicles): use company hook
The custom hook is used to obtain the company that was registered by the logged in user, this in order to obtain the employees of the same to assign them in the vehicle registration.
In case the user has not registered the company, he is redirected to the registration page.